Improve naming and slightly refactor.
