+++ client-cli/src/main/java/io/spine/examples/todolist/view/ViewOfTask.java
@@ -36,15 +36,15 @@ import static java.lang.System.lineSeparator;
 /**
  * An {@link EntityView} of a {@link TaskItem}.
  *
- * <p>Renders the task and provides actions for working with the task.
+ * <p>Renders a single task and provides actions for working with it.
  */
-final class TaskItemView extends EntityView<TaskId, TaskView> {
+final class ViewOfTask extends EntityView<TaskId, TaskView> {
 
     static final String DUE_DATE_VALUE = "Due date: ";
     static final String DESCRIPTION_VALUE = "Description: ";
     static final String PRIORITY_VALUE = "Priority: ";
 
-    TaskItemView(TaskId id) {
+    ViewOfTask(TaskId id) {
         super(id, "My task details");
     }
 
@@ -52,15 +52,12 @@ final class TaskItemView extends EntityView<TaskId, TaskView> {
     protected TaskView load(TaskId id) {
         //TODO:2017-07-19:dmytro.grankin: Allow to specify the source projection of task items.
         List<TaskView> tasks = getClient().taskViews();
-        Optional<TaskView> optionalTask = tasks.stream()
-                                               .filter(task -> task.getId()
-                                                                   .equals(id))
-                                               .findFirst();
-        if (optionalTask.isPresent()) {
-            return optionalTask.get();
-        }
-
-        throw newIllegalStateException("There is no task with ID `%s`.", id);
+        Optional<TaskView> task = tasks.stream()
+                                       .filter(view -> view.getId()
+                                                           .equals(id))
+                                       .findFirst();
+        return task
+                .orElseThrow(() -> newIllegalStateException("There is no task with ID `%s`.", id));
     }
 
     @Override
